+++ apps/docs/src/app/core/component-docs/message-box/examples/mobile-mode/message-box-mobile-example.component.ts
@@ -1,12 +1,17 @@
 import { Component } from '@angular/core';
+
 import { MessageBoxContent, MessageBoxService } from '@fundamental-ngx/core/message-box';
 
 @Component({
     selector: 'fd-message-box-mobile-example',
-    templateUrl: './message-box-mobile-example.component.html'
+    templateUrl: './message-box-mobile-example.component.html',
+    providers: [
+        // The MessageBoxService is already provided on the MessageBoxModule module.
+        // We do it at the component level here, due to the limitations of our example generation script.
+        MessageBoxService
+    ]
 })
 export class MessageBoxMobileExampleComponent {
-
     constructor(private _messageBoxService: MessageBoxService) {}
 
     open(): void {