+++ libs/core/src/lib/fundamental-ngx.module.ts
@@ -73,6 +73,7 @@ import { IllustratedMessageModule } from '@fundamental-ngx/core/illustrated-mess
 import { MenuModule, MenuMobileModule } from '@fundamental-ngx/core/menu';
 import { TabsModule } from '@fundamental-ngx/core/tabs';
 import { ResizableCardLayoutModule } from '@fundamental-ngx/core/resizable-card-layout';
+import { UploadCollectionModule } from '@fundamental-ngx/core/upload-collection';
 
 import { PaginationModule } from '@fundamental-ngx/core/pagination';
 import { SelectMobileModule, SelectModule } from '@fundamental-ngx/core/select';
@@ -82,6 +83,7 @@ import { WizardModule } from '@fundamental-ngx/core/wizard';
 import { SplitButtonModule } from '@fundamental-ngx/core/split-button';
 import { SideNavigationModule } from '@fundamental-ngx/core/side-navigation';
 import { MessagePageModule } from '@fundamental-ngx/core/message-page';
+import { MicroProcessFlowModule } from '@fundamental-ngx/core/micro-process-flow';
 import { TimelineModule } from '@fundamental-ngx/core/timeline';
 
 @NgModule({
@@ -178,7 +180,9 @@ import { TimelineModule } from '@fundamental-ngx/core/timeline';
         GridListModule,
         ResizableCardLayoutModule,
         MessagePageModule,
-        TimelineModule,
+        UploadCollectionModule,
+        MicroProcessFlowModule,
+        TimelineModule
     ],
     providers: []
 })