+++ libs/core/src/lib/micro-process-flow/injection-tokens.ts
@@ -1,0 +1,9 @@
+import { InjectionToken } from '@angular/core';
+import { Subject } from 'rxjs';
+
+export interface MicroProcessFlowComponentInterface {
+    canItemsReceiveFocus: Subject<boolean>;
+    setFocusedElementIndex(elm: HTMLElement): void;
+}
+
+export const MICRO_PROCESS_FLOW = new InjectionToken<MicroProcessFlowComponentInterface>('Micro process flow component dependency');