+++ libs/core/src/lib/utils/functions/index.ts
@@ -6,3 +6,4 @@ export * from './uuidv4-generator';
 export * from './parser-file-size';
 export * from './scroll';
 export * from './random-color-accent';
+export * from './clone-deep';