+++ libs/platform/src/lib/icon-tab-bar/public_api.ts
@@ -1,0 +1,2 @@
+export * from './icon-tab-bar.module';
+export * from './icon-tab-bar.component';